+++ src/components/Form/Form.tsx
@@ -1,26 +1,14 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import { SubmitHandler } from 'react-hook-form';
+import { COUNTRIES, SUPPORTED_FORMATS } from '../../constants/constants';
 import MyButton from '../UI/MyButton';
 import MyInput from '../UI/MyInput';
 import MySelect from '../UI/MySelect';
 import MyToggle from '../UI/MyToggle';
 import MyCheckbox from '../UI/MyCheckbox';
+import useValidation from './useFormWithValidation';
 import classes from './Form.module.scss';
-import { countriesType, FormCard, Gender } from '../../types/formTypes';
-import { useForm, SubmitHandler } from 'react-hook-form';
-import { COUNTRIES, SUPPORTED_FORMATS } from '../../constants/constants';
-interface FormProps {
-  addCard: (card: FormCard) => void;
-}
-
-export interface FormValues {
-  name: string;
-  surname: string;
-  gender: Gender;
-  date: Date | string;
-  country: countriesType;
-  dataProcessing: boolean;
-  file: FileList | null;
-}
+import { FormCard, FormValues } from '../../types/formTypes';
 
 const defaultValues: FormValues = {
   name: '',
@@ -31,19 +19,28 @@ const defaultValues: FormValues = {
   date: '',
   file: null,
 };
+interface FormProps {
+  addCard: (card: FormCard) => void;
+}
 
 const Form = ({ addCard }: FormProps) => {
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors },
-    clearErrors,
-  } = useForm<FormValues>({
-    defaultValues,
-  });
   const [logo, setLogo] = useState<null | string>(null);
   const [submitDisable, setSubmitDisable] = useState(true);
+  const {
+    handleSubmit,
+    reset,
+    errors,
+    clearErrors,
+    register,
+    registerValidation: {
+      countryRegister,
+      dataProcessingRegister,
+      dateRegister,
+      fileRegister,
+      nameRegister,
+      surnameRegister,
+    },
+  } = useValidation(defaultValues);
 
   const resetForm = () => {
     reset(defaultValues);
@@ -107,48 +104,6 @@ const Form = ({ addCard }: FormProps) => {
   const submitButtonDisabled = () => {
     return submitDisable || Object.values(errors).some((error) => Boolean(error));
   };
-
-  const nameRegister = register('name', {
-    required: 'Name is required',
-    minLength: {
-      value: 3,
-      message: 'Name should have at least 3 characters',
-    },
-  });
-  const surnameRegister = register('surname', {
-    required: 'Surname is required',
-    minLength: {
-      value: 3,
-      message: 'Surname should have at least 3 characters',
-    },
-  });
-
-  const fileRegister = register('file', {
-    required: 'You need to provide a file',
-    validate: {
-      fileType: (files) =>
-        (files && SUPPORTED_FORMATS.includes(files[0].type)) || 'Unsupported File Format',
-      fileSize: (files) => (files && files[0].size <= 1000000) || 'File size too large',
-    },
-  });
-
-  const dateRegister = register('date', {
-    valueAsDate: true,
-    required: 'Date is required',
-    validate: {
-      maxDate: (date) => date < new Date() || 'Pick correct birth date',
-    },
-  });
-
-  const countryRegister = register('country', {
-    validate: {
-      notDefault: (country) => country !== 'default' || 'Field required',
-    },
-  });
-
-  const dataProcessingRegister = register('dataProcessing', {
-    required: 'Data processing is required',
-  });
 
   return (
     <form
